Add exception, static files, js

+++ src/main/java/com/case_opener_game/case_opener/dto/bootstrap/BootstrapDTO.java
@@ -13,7 +13,7 @@ public class BootstrapDTO {
 
     private String game;
 
-    private List<Integer> multipliers;
+    private List<Double> multipliers;
 
     private String difficulty;
 }
